+++ commands/src/main/java/net/apartium/cocoabeans/commands/parsers/StringsParser.java
@@ -11,6 +11,7 @@
 package net.apartium.cocoabeans.commands.parsers;
 
 import net.apartium.cocoabeans.commands.CommandProcessingContext;
+import org.jetbrains.annotations.ApiStatus;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,8 +20,24 @@ import java.util.Set;
 
 public class StringsParser extends ArgumentParser<String> {
 
+    public static final String DEFAULT_KEYWORD = "strings";
+
+    /**
+     * Creates a new strings parser
+     * @param priority parser priority of which should be higher than others or lower
+     * @param keyword parser keyword
+     */
+    @ApiStatus.AvailableSince("0.0.36")
+    public StringsParser(int priority, String keyword) {
+        super(keyword, String.class, priority);
+    }
+
+    /**
+     * Creates a new strings parser
+     * @param priority parser priority of which should be higher than others or lower
+     */
     public StringsParser(int priority) {
-        super("strings", String.class, priority);
+        this(priority, DEFAULT_KEYWORD);
     }
 
     @Override