+++ commands/src/main/java/net/apartium/cocoabeans/commands/parsers/WithParserFactory.java
@@ -1,9 +1,9 @@
 package net.apartium.cocoabeans.commands.parsers;
 
 import net.apartium.cocoabeans.commands.CommandNode;
+import net.apartium.cocoabeans.reflect.ConstructorUtils;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
@@ -11,6 +11,7 @@ import java.lang.reflect.GenericDeclaration;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -28,7 +29,6 @@ public class WithParserFactory implements ParserFactory {
                 result.addAll(getArgumentParser(commandNode, withParser, obj));
 
             return result;
-
         }
 
         if (!(annotation instanceof WithParser withParser))
@@ -36,35 +36,84 @@ public class WithParserFactory implements ParserFactory {
 
 
         try {
-            Constructor<? extends ArgumentParser<?>>[] ctors = (Constructor<? extends ArgumentParser<?>>[]) withParser.value().getDeclaredConstructors();
-            return List.of(new ParserResult(newInstance((Constructor<ArgumentParser<?>>[]) ctors, withParser.priority()), obj instanceof Method ? Scope.VARIANT : Scope.CLASS));
+            ArgumentParser<?> argumentParser = newInstance(
+                    ConstructorUtils.getDeclaredConstructors(withParser.value()),
+                    withParser.priority(),
+                    withParser.keyword()
+            );
+
+            if (argumentParser == null)
+                return List.of();
+
+            return List.of(new ParserResult(
+                    argumentParser,
+                    obj instanceof Method ? Scope.VARIANT : Scope.CLASS
+            ));
         } catch (InstantiationException | IllegalAccessException | InvocationTargetException e) {
+            SharedSecrets.LOGGER.log(System.Logger.Level.WARNING, "Failed to create WithParser", e);
             return List.of();
         }
     }
 
-    private static <T> T newInstance(Constructor<T>[] ctors, int priority) throws InstantiationException, IllegalAccessException, InvocationTargetException {
-        Constructor<?> constructor = null;
+    private static <T extends ArgumentParser<?>> T newInstance(Collection<Constructor<T>> constructors, int priority, String keyword) throws InstantiationException, IllegalAccessException, InvocationTargetException {
+        Constructor<T> matchingConstructor = null;
 
-        if (ctors.length > 1) {
-            for (Constructor<T> ctor : ctors) {
-                if (ctor.getParameterCount() == 1 && ctor.getParameterTypes()[0].equals(int.class))
-                    constructor = ctor;
+        boolean keywordBlank = keyword.isBlank();
+        Class<?>[] paramTypes = null;
+        for (Constructor<T> constructor : constructors) {
+            int paramCount = constructor.getParameterCount();
+
+            if (paramCount != 1 && paramCount != 2)
+                continue;
+
+            paramTypes = constructor.getParameterTypes();
+
+            if (keywordBlank && paramCount == 1 && paramTypes[0].equals(int.class)) {
+                matchingConstructor = constructor;
+                break;
+            }
+
+            if (!keywordBlank && paramCount == 2) {
+                if (paramTypes[0].equals(int.class) && paramTypes[1].equals(String.class)) {
+                    matchingConstructor = constructor;
+                    break;
+                }
+
+                if (paramTypes[0].equals(String.class) && paramTypes[1].equals(int.class)) {
+                    matchingConstructor = constructor;
+                    break;
+                }
             }
         }
 
-        if (constructor == null)
-            constructor = ctors[0];
+        if (matchingConstructor == null) {
+            if (keywordBlank) {
+                SharedSecrets.LOGGER.log(System.Logger.Level.WARNING,
+                        "No constructor found for WithParser with priority {}",
+                        priority
+                );
+            } else {
+                SharedSecrets.LOGGER.log(System.Logger.Level.WARNING,
+                        "No constructor found for WithParser with priority {} and keyword {}",
+                        priority, keyword
+                );
+            }
+
+            return null;
+        }
 
         Object[] params;
-        if (constructor.getParameterCount() == 1) {
-            params = new Object[] {priority};
-        } else {
-            if (priority > 0)
-                SharedSecrets.LOGGER.log(System.Logger.Level.WARNING, "Registered parser {} with priority, but it doesn't support it", constructor.getDeclaringClass().getName());
 
-            params = new Object[0];
+        if (!keywordBlank) {
+            if (paramTypes[0] == int.class)
+                params = new Object[] { priority, keyword };
+            else
+                params = new Object[] { keyword, priority };
+        } else {
+            params = new Object[] { priority };
         }
-        return (T) constructor.newInstance(params);
+
+        return matchingConstructor.newInstance(params);
     }
+
 }